6858429: javap classfile library a minor bug

+++ langtools/src/share/classes/com/sun/tools/classfile/ClassWriter.java
@@ -435,7 +435,7 @@ public class ClassWriter {
         }
 
         public Void visitLocalVariableTypeTable(LocalVariableTypeTable_attribute attr, ClassOutputStream out) {
-            out.writeByte(attr.local_variable_table.length);
+            out.writeShort(attr.local_variable_table.length);
             for (LocalVariableTypeTable_attribute.Entry e: attr.local_variable_table)
                 writeLocalVariableTypeTableEntry(e, out);
             return null;
